update dgrasp raisimgymtorch files

+++ raisimGymTorch/raisimGymTorch/env/RaisimGymVecEnv.py
@@ -20,7 +20,7 @@ class RaisimGymVecEnv:
         self.wrapper = impl
         self.num_obs = self.wrapper.getObDim()
         self.num_acts = self.wrapper.getActionDim()
-        self._observation = np.zeros([self.num_envs, self.num_obs], dtype=np.float64)
+        self._observation = np.zeros([self.num_envs, self.num_obs], dtype=np.float32)
         self.obs_rms = RunningMeanStd(shape=[self.num_envs, self.num_obs])
         self._reward = np.zeros(self.num_envs, dtype=np.float32)
         self._done = np.zeros(self.num_envs, dtype=np.bool)
